Add an auto discovery plugin for mixins
This will automatically find mixins in the mixin plugin, not requiring
the manual configuration. This is less friction, especially for newer
users. Usually this isn't super advisable, or needs more configuration
to avoid server/client mixins being incorrectly applied, but with the
target demographic of mostly client mods, I think this is a worthwhile
change.